added back some auth

+++ flask/resources/articles.py
@@ -8,8 +8,8 @@ article = Blueprint('articles', 'article')
 # index route
 @article.route('/<topicId>/', methods=['GET'])
 def index_articles(topicId):
-    #if not current_user.is_authenticated: # Checks if user is logged in
-        #return jsonify(data={}, status={'code': 401, 'message': 'You must be logged in to view articles'})
+    if not current_user.is_authenticated: # Checks if user is logged in
+        return jsonify(data={}, status={'code': 401, 'message': 'You must be logged in to view articles'})
     try:
         all_articles = [model_to_dict(a) for a in models.Article.select().where(models.Article.topic_id == topicId)]
         return jsonify(data=all_articles, status={'code': 200, 'message': 'Success'})
@@ -20,8 +20,8 @@ def index_articles(topicId):
 # show route
 @article.route('/show/<articleId>/',methods=['GET'])
 def show_article(articleId):
-    #if not current_user.is_authenticated: # Checks if user is logged in
-        #return jsonify(data={}, status={'code': 401, 'message': 'You must be logged in to view this article'})
+    if not current_user.is_authenticated: # Checks if user is logged in
+        return jsonify(data={}, status={'code': 401, 'message': 'You must be logged in to view this article'})
     try:
         found_article = models.Article.get(id=articleId)
         article_dict = model_to_dict(found_article)
@@ -38,10 +38,10 @@ def create_article(topicId):
     print(topicId)
     topic = models.Topic.get_by_id(topicId)
     topic_dict = model_to_dict(topic)
-    #if not current_user.is_authenticated:
-        #return jsonify(data={}, status={'code': 401, 'message': 'You must be logged in to save an article'})
-    #if topic_dict.user.id is not current_user.id: 
-            #return jsonify(data={}, status={'code': 401, 'message': 'You can only add articles to your own topics'})
+    if not current_user.is_authenticated:
+        return jsonify(data={}, status={'code': 401, 'message': 'You must be logged in to save an article'})
+    # if topic_dict.user.id is not current_user.id: 
+    #         return jsonify(data={}, status={'code': 401, 'message': 'You can only add articles to your own topics'})
     payload = request.get_json()
     payload['topic'] = topicId
     created_article = models.Article.create(**payload)
@@ -53,10 +53,11 @@ def create_article(topicId):
 # delete route
 @article.route('/<articleId>/', methods=['DELETE'])
 def delete_article(articleId):
+    if not current_user.is_authenticated:
+            return jsonify(data={}, status={'code': 401, 'message': 'You must be logged in to delete an article'})
     try:
         article_to_delete = models.Article.get(id=articleId)
-        #if not current_user.is_authenticated:
-            #return jsonify(data={}, status={'code': 401, 'message': 'You must be logged in to save an article'})
+        
         #if article_to_delete.topic.user.id is not current_user.id: 
             #return jsonify(data={}, status={'code': 401, 'message': 'You can only delete your own articles'})
         article_to_delete.delete_instance()
@@ -71,8 +72,8 @@ def update_article(articleId):
         payload = request.get_json()
         article = models.Article.get_by_id(articleId)
         article_dict = model_to_dict(article)
-        #if not current_user.is_authenticated:
-            #return jsonify(data={}, status={'code': 401, 'message': 'You must be logged in to save an article'})
+        if not current_user.is_authenticated:
+            return jsonify(data={}, status={'code': 401, 'message': 'You must be logged in to update an article'})
         #if article_dict.topic.user.id is not current_user.id: 
             #return jsonify(data={}, status={'code': 401, 'message': 'You can only delete your own articles'})
         updated_article = models.Article.update(
